+++ gdal-src/src/lib.rs
@@ -1,0 +1,2 @@
+extern crate libsqlite3_sys;
+extern crate proj_sys;